Create wiki15.md
##Diversity and Inclusion##

The main things I got out of this reading is that there is a massive lack of diversity in the tech field even though diversity is something desired in the workplace. Many factors have come into play here, but one of the ones when it comes to gender is video games and how they have mainly been marketed towards men, causing a decline in interest when it comes to women with technology and computers. This created a male dominated field, which made it unfriendly and intimidating for women who may have otherwise been interested in pursuing these subjects. The statistics shown to us also show that the tech companies are even currently very largely dominated by white men. Despite that, tech companies want more diversity in their companies as diverse teams are smarter and more creative. There are more people with different life experiences and ways of thinking, which enables solutions and ideas that would likely not be thought of otherwise. Diverse companies also perform better and can serve a more diverse user base. Not to mention, diversity and inclusion is the right thing for companies to strive for. 